Final Update
This is for me, A finished project, some features may be added in the future, but I don't think I'll ever touch that project again. The main changes are related to the UI/UX but some "back-end" features and or bug have been added & removed. Such as the text mode, altought not finished as much as i'd hope to. There's also the save / load. A customize How-To window that will now pop everytime you launch the program (Will probably work around this to make it pop only the first time later).

As you can guess the code is also a lot cleaner, there's been a whole new revamp of the codebases, with the uses of customize libaries, those libraries come as is, with the program but one could use for is own project. That's said, I have also implemented and went for a more object oriented programming, which support all following object  :

Brush, Log, ColorTable, TextPropreties as well as some other object related. 

I've also decided to let include the ColorFormatter which is in Java, that allows the conversion of a data set of color that I found online into the more usable .CSV file for the ColorTable object.


Thanks for reading, T1WiLLi. 
Have fun !